#7 0809 14:37 dev=>main CD

See merge request s11-webmobile2-sub2/S11P12A306!103

+++ spring-dekku/src/main/java/dekku/spring_dekku/domain/member/jwt/CookieUtil.java
@@ -3,11 +3,11 @@ package dekku.spring_dekku.domain.member.jwt;
 import jakarta.servlet.http.Cookie;
 
 public class CookieUtil {
-    public static Cookie createCookie(String key, String value, Integer expiredS) {
+    public static Cookie createCookie(String key, String value, Long expiredS) {
         Cookie cookie = new Cookie(key, value);
         cookie.setHttpOnly(true);
         cookie.setPath("/");
-        cookie.setMaxAge(expiredS);
+        cookie.setMaxAge(Math.toIntExact(expiredS));
         return cookie;
     }
 }
